pyproject.toml
[tool.poetry]
name = "doors-bot"
version = "0.1.0"
description = ""
authors = ["Almaz Ilaletdinov <a.ilaletdinov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "<4.0.0"
pyTelegramBotAPI = "^4.2.2"
django-environ = "^0.8.1"
django-split-settings = "^1.1.0"
django-debug-toolbar = "^3.2.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
drf-yasg = "^1.20.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-eradicate = "^1.2.0"
flake8-isort = "^4.1.1"
flake8-print = "^4.0.0"
isort = "^5.10.1"
flake8-docstrings = "^1.6.0"
pytest = "^6.2.5"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
